update: well
- Добавлен класс DominoProtectorSource для объектов, которые могут защищать домино (на данный момент это фонарь и колодец). У каждого объекта можно настроить свою вероятность защиты.
