+++ Assets/Prefabs/Buildings/HunterHouse.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2316e2bfd40f7ac40b8bc5aca60088f1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
