+++ Assets/Prefabs/Settlement.prefab
@@ -78,7 +78,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: 7482667652216324306, guid: 311925a002f4447b3a28927169b83ea6, type: 3}
-  m_Color: {r: 1, g: 0.504717, b: 0.51478374, a: 1}
+  m_Color: {r: 0.043211088, g: 0.4490565, b: 0.28752694, a: 0.5058824}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
