+++ Assets/Prefabs/UI(OnlyForCanvas)/ItemsAndDomino/HunterHouseUI.prefab
@@ -13,7 +13,7 @@ GameObject:
   - component: {fileID: 6758885814436890381}
   - component: {fileID: 4116706721653000829}
   m_Layer: 5
-  m_Name: HouseUI
+  m_Name: HunterHouseUI
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -67,7 +67,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: -7411974349392773033, guid: 1ab8986a32a9ed047ad6c49500857e5d, type: 3}
+  m_Sprite: {fileID: -8589871959191218572, guid: b5baa6f447d6e5640834a7a3ab576df2, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -89,7 +89,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2ec269458851ab84dbe08e4445763c7a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemId: house
+  itemId: hunterhouse
   blurImage: {fileID: 5900784217510290946}
 --- !u!1001 &8170348653032184749
 PrefabInstance:
@@ -101,7 +101,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 2342191990407282095, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_Name
-      value: HouseDark
+      value: HunterDark
       objectReference: {fileID: 0}
     - target: {fileID: 2342191990407282095, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_IsActive
@@ -110,7 +110,7 @@ PrefabInstance:
     - target: {fileID: 6758885814436890381, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_Sprite
       value: 
-      objectReference: {fileID: -7411974349392773033, guid: 1ab8986a32a9ed047ad6c49500857e5d, type: 3}
+      objectReference: {fileID: -8589871959191218572, guid: b5baa6f447d6e5640834a7a3ab576df2, type: 3}
     - target: {fileID: 8386460687464962531, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
@@ -145,7 +145,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 8386460687464962531, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_LocalScale.x
-      value: 1
+      value: 0.9989909
       objectReference: {fileID: 0}
     - target: {fileID: 8386460687464962531, guid: bb47d20ceb7b5e34d81e8939710d4555, type: 3}
       propertyPath: m_LocalScale.y
