+++ Assets/Prefabs/UI(OnlyForCanvas)/ItemsAndDomino/HunterHouseUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b0bf4a4c6a8dcc848a965c67775a3f0b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
