+++ Assets/Resources/Data/Items/HunterHouse.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b1c0cf2eb747d24fb4ca1a1c524fadb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
