+++ Assets/Resources/ItemManager.asset
@@ -24,3 +24,4 @@ MonoBehaviour:
   - {fileID: 11400000, guid: c951063245bc19549b0a2d14e04e7ddc, type: 2}
   - {fileID: 11400000, guid: c2cfb5a29bc32264b97e7a28e57ef873, type: 2}
   - {fileID: 11400000, guid: 3aef8fefaf27009409634d6173012693, type: 2}
+  - {fileID: 11400000, guid: 5b1c0cf2eb747d24fb4ca1a1c524fadb, type: 2}
