+++ Assets/Scripts/Buildings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d26add2dc24f4364cb72f6abbc6f2fb9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
