+++ Assets/Scripts/Items/Lantern.cs.meta
